- Hook up probabilistic downsampling functionality. - Refactor probabilistic utils.
